- [BE][FE] 마이데이터 큰 금액 지원 및 회원가입/자산연동 회귀 보강

See merge request s14-fintech-finance-sub1/S14P21C103!181

+++ backend/src/main/java/io/ssafy/p/j14c103/homerun/api/service/account/UserSsafyAccountSyncService.java
@@ -231,12 +231,12 @@ public class UserSsafyAccountSyncService {
         return LocalDateTime.of(localDate, localTime);
     }
 
-    private Integer toAmount(final Object value) {
+    private Long toAmount(final Object value) {
         final String text = toText(value);
         if (text == null) {
-            return 0;
+            return 0L;
         }
-        return Integer.parseInt(text);
+        return Long.parseLong(text);
     }
 
     private String toText(final Object value) {
